hlsl: Use unique register namespaces for resources
Instead of going with the SPIRV-Cross default of mapping a SPIR-V binding
N to HLSL register bN (buffers), tN + sN (combined image samplers), or uN
(storage buffers and images), request it to use bX, sY, tY, uZ (where X, Y, Z
runs independently from 0) registers in the generated HLSL source.

We then store the N -> X, N -> (Y, Y), or N -> Z mappings in the QShader's
nativeResourceBindingMap. The D3D11 backend of QRhi is expected to use
this map instead of just assuming that the register binding is the same
s the SPIR-V binding. (although it will still follow that assumption for old
QShaders that do not have this map for HLSL, so we stay compatible with
any existing .qsb file).

Add a manyresources.frag test shader that can be used to verify this (qsb
-d prints the native resource binding map as well). Our single SPIR-V binding
list (1, 2, 3, 4, 5, 10, 11, 12) should end up with the following mapping
table (where 1, 2, 3 are combined image samplers, 4 and 5 are storage images,
10 is a storage buffer, 11 and 12 are uniform buffers):

1 -> [0, 0] // t0, s0
2 -> [1, 1] // t1, s1
3 -> [2, 2] // t2, s2
4 -> [1, -1] // u1
5 -> [2, -1] // u2
10 -> [0, -1] // u0
11 -> [0, -1] // b0
12 -> [1, -1] // b1

Task-number: QTBUG-82472
Change-Id: I745f825269ac115b848ac306bad15cf0f2667ddf
Reviewed-by: Christian Strømme <christian.stromme@qt.io>