Change qshaderbaker.h into a private header
We have no choice given that QShader (in QtGui) is private. Including
a private header from a public header is Not a Good Thing. One needs
gui-private anyways to be able to do anything with the resulting QShader.
Having to pull in shadertools-private as well, instead of just
shadertools, does not make much of a difference then.

It remains to be seen how the shader (and QRhi) story evolves long-term,
but for now (6.0) we are stuck with it all being private.

Change-Id: I00eba277b1ffcc01e5e9f2c1c41019f997b7b162
Reviewed-by: Christian Strømme <christian.stromme@qt.io>