Strip all SPIRV debug in qsb and add full debug info option
Once the SPIRV-Cross translation is done there is no need for variable
names and such, strip them all. This gives us smaller .qsb files.

At the same time, add an option to opt in for full debug info, which
involves embedding the source into the SPIR-V binary.  So now when
passing -g to qsb, one can see the full original source in RenderDoc
in the shader views, and the new Vulkan shader debugging in RenderDoc 1.9
can also switch to the actual source instead of just doing SPIR-V disassembly.

In qsb -g is also supported in combination with -c (aka --fxc), meaning when
fxc is invoked. This will trigger passing /Od /Zi to fxc. This is again
extremely handy with RenderDoc because one can now see and step through
the the original (SPIRV-Cross generated) HLSL source, also in the vertex
and pixel debuggers.

The downside of -g is that the generated .qsb files are significantly
larger and also disables any optimizations.

Task-number: QTBUG-85949
Change-Id: I6dee5c199cdecc7235b3eba62e6be2a57ec099fd
Reviewed-by: Christian Strømme <christian.stromme@qt.io>