qsb: Add an option to invoke spirv-opt to get optimized SPIR-V
Getting SPIRV-Tools bundled is too complicated at this point. So
instead, add an option to qsb to invoke spirv-opt, assuming one has
a Vulkan SDK installed. This will run spirv-opt -O (optimize for
performance) and include the resulting SPIR-V binary in the .qsb file
instead of the original. So this is in practice a post processing step
similar to invoking fxc or other external tools. It does not allow to
get optimized SPIR-V when using QShaderBaker at run time, but that's
something we have to live with for now.

Also simplify the fxc and xcrun metal/metallib branches as much as
possible so that we have more compact code.

Change-Id: Iec349d9d63bb3627212390ce393770b5edfc366a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>