qsb: Make failing to launch external processes non-fatal
Most relevant with -O: when spirv-opt could not be run, it failed
the entire qsb run previously. This is not ideal, considering that
one may not have control over the environment in which CMake or
other scripts invoking qsb are used, and that these post-processing
steps are optional.

Make all post-processing steps non-fatal in the sense that if the
external tool cannot be launched (fxc, spirv-opt, XCode tools, etc.
is missing), just skip that step.

Pick-to: 6.1
Change-Id: Ie2f2d1cc1f71084a136c25e8641e927c22680df4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>