qsb: Do mkpath before writing files
So doing -o x/y/z.qsb will work even when the directories
x and y do not exist yet.

In theory this should not be needed when using the CMake
integration, but apparently the output subdirectories do
not always get created as expected by the build system. So
add this into the tool.

Also add the Truncate flag when opening a QFile for writing.

Pick-to: 6.1
Change-Id: Ie1fe6e4ea043508f60e241f91fe4c7133ef882a3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>