qsb: Adjust silent flag's effects
Apply the !silent guard to external tool invocations.
The callers of runProcess already guard their qWarnings
like this, now add it to the function itself as well.
This becomes relevant in particular when running qsb via
CMake, and tools like spirv-opt are not present.

Adjust the help text as well, it was not quite correct in
the first place.

Change-Id: I35ca14a02824d4023928c85148419b941f1b33b9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 919cef4874ceeaa87af723e08fe99ddbe1c6751f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>