Add qsb option to replace a given shader with data read from a file
Say we have a fragment shader for OpenGL ES that uses
samplerExternalOES. This cannot be sent through glsland/SPIRV-Cross.

Rather, what we can do is to process a Vulkanized version of the
fragment shader without the GLES-specific constructs. Once a .qsb file
is available, we can replace the contents for the GLSL shader with
the contents of a file (the fragment shader written specifically for
GLES, using samplerExternalOES and co.)

While we are at it, the argument handling for extraction is updated to
be compatible between the two functions (extract, replace):

qsb -x glsl,100es -o shader.frag shaderpack.qsb
writes the GLSL ES 100 shader to shader.frag

qsb -r glsl,100es,myspecialshader.frag shaderpack.qsb
replaces the GLSL ES 100 shader in shaderpack.qsb with the contents of
myspecialshader.frag. Multiple -r arguments are supported as well.

With CMake, the option to specify replacements with the help of magic
separator characters is there in the FILES list:

FILES
  "wobble.frag@glsl,100es,custom_gles_shader.frag@spirv,100,custom_spirv_binary.spv"

Here wobble.frag.qsb is generated normally, but the normal qsb run is
followed up by another call to qsb, this time with two -r arguments to
replace two of the shaders in the .qsb file with the specified content.

Change-Id: Ib12c33a9d19958a1a7b2364b406ec654abe12ac0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 01aef9842f5f6e0d0d5497c0d63da629a70da63a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>