qsb: Improve shader replacement functionality
With CMake, use absolute source file paths to support shadow builds.

Exit with non-zero value when extract or replace fails, this is again
useful for the CMake integration, because if the source file does not
exist, the build should be aborted.

Fix the case of multiple replacements per qsb invocation.

Pick-to: 6.2
Change-Id: I1a82f43a5491ce265560a033de1f5bb8b6d5c937
Reviewed-by: Lars Knoll <lars.knoll@qt.io>