Conan: Inherit recipe class from QtLeafModule for common functionality
The qt-conan-common package implements a base class for Qt leaf
module recipes. The build steps in leaf modules are mostly identical
so it makes sense to put those in the base class.

Dependencies are read by the base class from the 'dependencies.yaml'
which is the same file the CI system uses.

Change-Id: I2f1473d31d999ef7152db1b7afcbba746984c447
Reviewed-by: Toni Saario <toni.saario@qt.io>