Change PREFIX and FILES path handling to match add_resources
Compare the following:

qt6_add_resources(target "stuff"
    PREFIX "/some/prefix"
    FILES "subdir/texture.frag"
)

qt6_add_shaders(target "other_stuff"
    PREFIX "/some/prefix"
    FILES "subdir/texture.frag"
)

Based on the expectation of API consistency, one could assume that
result is a :/some/prefix/subdir/texture.frag[.qsb] file available in
in the resource system.

This is not the case. add_shaders strips the path, so the result is
:/some/prefix/texture.frag.qsb. Thus if someone wants the "subdir" in
there, it needs to be specified in the PREFIX. This is what current
users, such as qtdeclarative or qtquick3d, do, along the following
lines:

qt6_add_shaders(target "more_stuff"
    PREFIX "/some/prefix/subdir"
    FILES "subdir/texture.frag"
)

Note how subdir is baked in to the PREFIX to counter add_shaders'
behavior.

This is both inconsistent, and also limiting. (what if there were
different paths for the files, e.g. subdir1/x.vert, subdir2/y.frag,
etc., and we would want to keep the subdir1, subdir2, ... - right now
this can only be achieved by either specifying OUTPUTS, or creating
separate qt6_add_shaders() calls for each file)

Speaking of OUTPUTS, the stripping of the path in the FILES entry
comes from a time when OUTPUTS did not exist. Thus specifying an
absolute path in FILES made the add_shaders function unusable. Whereas
in the rcc world this is solved by aliases. Today OUTPUTS solves this
problem, because one can do:

qt6_add_shaders(target "even_more_stuff"
    PREFIX "/some/prefix"
    FILES "c:/some/where/texture.frag"
    OUTPUTS "yay/texture.frag"
)

which generates :/some/prefix/yay/texture.frag.qsb. In fact the
current behavior with FILES is inconsistent with OUTPUTS: when OUTPUTS
is present, the processing that would be done on FILES[i] is not done
on OUTPUTS[i] at all, rather, OUTPUTS[i] is taken as-is.

Therefore, start using the FILES entries as-is, dropping the
get_filename_component-based adjustments.

Also introduces BASE for consistency. This matches add_resources:

qt6_add_shaders(target "stuff_with_base"
    PREFIX "/some/prefix"
    BASE "../stuff"
    FILES "../stuff/shader.vert")

this gives :/some/prefix/shader.vert, without having to specify
"shader.vert" manually in OUTPUTS.

[ChangeLog] The qt6_add_shaders CMake function has been changed to match
qt6_add_resources when it comes to the handling of PREFIX, BASE, and
FILES. qt6_add_shaders no longer strips the directories from the entries in
the FILES list. This means that invocations that previously catered for
the special behavior by listing the necessary extra directories in PREFIX
must now be changed to not do that anymore.

Change-Id: I605dd76f60ec668cde4d3541257727b6f005cd10
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit b4394f25f7614668d2966373346584d4108bab47)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>