qsb: Make error messages visible from Windows GUI applications
When running qsb from a GUI program on Windows, like Qt Creator or
Visual Studio, no qsb error messages would appear in the build output.
This was, because qWarning was used for error messages.  But qWarning
uses the debugger API on Windows if no console is attached.

Replace qWarning with a function that wraps fprintf(stderr, ...).

Pick-to: 6.2
Fixes: QTBUG-96524
Change-Id: I9b92df1f608d6ee147d710089e75ef148e3697ba
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>