qsb: Add an option to erase something from the shader pack
...while clarifying the capabilities of the remove command as well.

Make it clear that replace (-r) can also add a shader for a key that was
not present at all before.

Add an erase (-e) option that, similarly to extract and add/replace,
allwos doing the missing operation: removing a shader for a given key,
and then writing out the result to the same .qsb file.

Change-Id: Ie477ec67f0b5014514143e0569d3c692ef0c26ae
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>