Update commercial license headers
Updated header.COMM to the files in tqtc-qtshadertools.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4930
Change-Id: I9c77941d50305ca811b92644199b49697d62ce17
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>