CMake: Fix Xcode common dependency issues for shader files
When using the Xcode generator (usually for iOS projects),
make sure to create a custom target that will drive the processing
of shaders. This target will be used as a common dependency for
the resource processing target as well as the ide other_files target.

This should fix errors like

  The custom command generating

    foo.frag.qsb

  is attached to multiple targets:

    foo_other_files
    foo_resources_1

  but none of these is a common dependency of the other(s).  This is
  not allowed by the Xcode "new build system".

This is similar to 5b0e765ab0dddba86662925cb44aeac748a286b7 in
qttools.

Pick-to: 6.2 6.3 6.4
Fixes: QTBUG-103723
Change-Id: I421e5f477c215115bc4b3679efca5e48d7fcd12a
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>