Update glslang
...from 11.1.0 (early 2021) to 11.10.0 (06/2022).

Some big endian patches have been upstreamed by the contributors so
those patches are removed since they are part of the upstream content.

The INTEGRITY hacks are split into two: one is the renaming of Sh*
functions, the other is the namespace change. The latter is different
from before, now we rename the entire glslang namespace to qglslang with
a search-and-replace. Hopefully this is still sufficient.

Change-Id: Icf0d1453d517cc5a36ac13bf83c2d30476a41c9c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>