Add enablers for Metal tessellation
qsb now has a way to say that a vertex shader is going to be used with
tessellation. For non-Metal this is ignored. With Metal this generates
the three variants.

Opting in is done by passing --msltess to qsb (so e.g.
qsb --msl 12 --msltess shader.vert) or by specifying the three variants
(UInt32IndexedVertexAsComputeShader etc.) when using QShaderBaker
at run time.

For vertex, tess.control, and tess.eval it now also stores the extra
buffer indices the QRhi Metal backend will need to know about.

And store a bunch of other things as well.

For tess.control one will need to provide the tessellation mode
(triangle/quad), there is no way to know this without user input
because in GLSL this is in the tess.eval. shader.

For tess.eval. one will need to provide the tess.control stage's
output vertex ocunt, there is no way to know this without user
input because in GLSL this is declared in the tess.control shader.

Change-Id: I749279e9fd1782247de9c589f5cd6b39b6d139a3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>