qsb: CMake: Enable PRECOMPILE on macOS
Pass -t on to qsb whenever PRECOMPILE is specified in qt_add_shaders(),
which will in turn attempt to invoke metal/metallib in order to ship
the intermediate format, not MSL source in the .qsb files.

This is expected to be stable enough now that we do not need to consider
older XCode SDKs.

iOS is not affected for now since that would need further testing.

Change-Id: I9a04daa051a81b36f265334bad9b6ef6b765fb1c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>