Revert "qsb: CMake: Enable PRECOMPILE on macOS"
We need to pass more arguments to the "metal" tool, like specifying
-std=macos-metal1.2, in order to make it robust, apparently.
Otherwise interesting versioning problems arise when building
on a newer macOS and SDK than the deployment target.

For now we'll just refrain from making things run fast on macOS and
rather stick with the old behavior of shipping Metal shader source code
and compiling it to the intermediate format at run time.

This reverts commit d213f969dd62ca019dcca9a32d82a3bf555013da.

Change-Id: Id8943efba7a4c50d21c601464598d4e997f229ed
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>