qsb: Make -t specify the language version
Also add a new option to choose iOS. The end goal is to be able
to generate an argument in the form of

-std=<platform>-metal<major>.<minor>

where platform is either "macos" or "ios".

This attempts to avoid problems where one calls qsb with -t, and
the resulting .qsb files contain compiled code for whatever version
the SDK decides to use, and then the application is deployed to an
older macOS that rejects the metallib binaries with that newer version.

Change-Id: Id78bba22075a7a003f3bcb22a102690a70794ebc
Reviewed-by: Christian Strømme <christian.stromme@qt.io>