Fix incorrect parsing in the rewriter
What we inherit from the scenegraph from Qt 5 has an
obvious bug, where tokens starting with 'v' get that
letter cut off. This does not affect the behavior of
the batching-rewriter, since that only really cares
about looking for "void" followed by "main", where
void has its own token and so the problem never
becomes visible, but is incorrect once we start to
use the token stream for other purposes.

While we are at it, remove the unused enums and variables.

Change-Id: Ie4d5168ccecaac61790e02f96294566637a44894
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>