Add multiview support for SPV/GLSL/HLSL
OpenGL is the interesting one, this needs a --view-count 2 (or
some other >= 2 value) argument for qsb in order to inject the
layout(num_views = 2) in; statement. See example vertex shader
added to tests/playground.

SPIR-V should work as-is. HLSL needs >= 6.1 (and D3D12 presumably
on the engine side) so specify --hlsl 61 or higher.

QShaderBaker and the CMake integration are updated with the new
view count parameter as well.

Metal is currently unknown and left as a future exercise.

While we are at it, the qsb doc page gets a few updates so that
the included qsb output reflects what is printed with dev atm.

Change-Id: I9b79464a925d5bed15e3046b6e0b90f2458ea33e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>