qsb: Call dxc instead of fxc for SM 6.0+
Relevant when wanting to (having to) use a recent HLSL version
for D3D12, e.g. because we need 6.1 to get SV_ViewID. fxc would not
compile that. The result is only compatible with the D3D12 backend
because the D3D11 backend only looks for DXBC, not DXIL. This is
probably fine for now, because one should know that when asking for
something larger than 5.0, it is something that is relevant with
D3D12 only.

This provides the simpler part of the associated task. Doing the
same at runtime (when compiling from HLSL, which happens anytime
one does not specify -c for qsb or uses QShaderBaker for dynamically
generated shaders) is a different story.

Task-number: QTBUG-114773
Change-Id: I0123742de2a2ea8a414a1c1709ed137966a95c67
Reviewed-by: Andy Nichols <andy.nichols@qt.io>