Update SPIRV-Cross to sdk-1.3.261.0
Includes the INTEGRITY and QNX patches, but it needs to be redone
since the existing two .patch files do not apply. Hence a new single
.patch now.

There is a behavior change around HLSL, it seems. One autotest case's
shader is adjusted for this since the existing one is not translatable
anymore to HLSL for some reason. (buffer packing rules)

Task-number: QTBUG-116484
Change-Id: I55f4480c8dbb572eea7b4f3638d16abb691539fd
Reviewed-by: Christian Strømme <christian.stromme@qt.io>