hlsl: Force storage buffer mapping to UAVs even for readonly
Read-only storage blocks generating an SRV (and ByteAddressBuffer)
instead of an UAV (and RWByteAddressBuffer) is not great, given
nothing is prepared for it. Both the register mapping logic gets
confused (as it expects 'u' not 't'), and the QRhi backends are
also not prepared for anything else but UAV+RWByteAddressBuffer.

Setting the flag also makes buffers consistent with images.
Storage images translate to UAV + RWTexture2D by default even
when non-writable.

Change-Id: Icafd5328ae7f27236ad9271baaabb0ba8156feb6
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>