Add depfile support to qsb
This adds the new option '--depfile' to qsb tool. The tool now supports
generating of gcc depfiles based on the '#include' statements found in
the shader files. The generated depfile it then passed to the CMake
add_custom_command command that calls qsb to control changes in the
shader includes.

Change-Id: I6330fbaaef35cb294f1e3ce7a3054f836fe986db
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>