glslang: Fix/Workaround crash on Android (arm64-v8a) in debug mode
When compiling shaders (GLSL->SPIRV) at runtime on Android (arm64-v8a)
when targeting debug mode, we get a crash in the STL/C++ runtime code
just based on the particular usage patched.  By changing how this
operation is performed, we can work around the triggered bug. This
bug has already been reported upstream, and this workaround comes from
that bug report. The fix was not applied upstream because they claim it
is a bug in the Android NDK/runtime, and this is just a workaround, not
a fix. That is fine for us though, we have many workaround patches.

Pick-to: 6.8
Change-Id: I44bdd148a85e3ede0924404aecfd7f0feeb16613
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>