CMake: Make it depend on the input absolute filename too
Does not make a differences when ORIGINAL_FILES is not specfied.
In that case the two variables are identical (the same filename).

When it is, however, doing DEPENDS ${original_file} is insufficient,
and we want to depend both on the input and the original file (on
which the input is based, generated by some other tool, etc. that
we do not know, but it needs to exist, and the qsb call must not
be made while it doesn't)

This avoids subtle build issues due to parallelism, where e.g. a
tool generated shader sources gets invoked after, or in parallel,
with qsb in some cases, which if of course wrong.

Change-Id: I402ec39b65889f83297902bc7076a02b8f7f16bf
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>